i think this is what you meant
